+++ docs/changes.rst
@@ -20,6 +20,7 @@ Not yet released.
 * Browse mode can now be navigated using keyboard, see :ref:`keyboard`.
 * :http:get:`/api/components/(string:project)/(string:component)/credits/` and :http:get:`/api/projects/(string:project)/credits/` API endpoints for components and projects.
 * :ref:`glossary-terminology` entries in Glossary can now only be created by users with :guilabel:`Add glossary terminology` permission.
+* :ref:`check-python-brace-format` detects extra curly braces.
 
 **Bug fixes**